refactor(e2e)
